:recycle: move keycloak & redis token IO into an infrastructure layer
Rename token_exchange_service -> infrastructure/keycloak and
redis_token_store -> infrastructure/redis, and lift ensure_offline_token
into a use-case, so token IO lives below the service layer.
