+++ mcr-core/mcr_meeting/app/infrastructure/keycloak.py
@@ -4,9 +4,6 @@ from keycloak import KeycloakOpenID
 from loguru import logger
 
 from mcr_meeting.app.configs.base import KeycloakExchangeSettings
-from mcr_meeting.app.services.redis_token_store import (
-    save_refresh_token,
-)
 
 
 @dataclass(frozen=True)
@@ -64,15 +61,3 @@ def refresh_access_token(refresh_token: str) -> TokenRefreshResult:
     return TokenRefreshResult(
         access_token=result["access_token"], rotated_refresh=rotated
     )
-
-
-def ensure_offline_token(user_sub: str, access_token: str | None) -> None:
-    if access_token is None or _keycloak is None:
-        return
-
-    try:
-        refresh_token = exchange_token_for_offline(access_token)
-        if refresh_token:
-            save_refresh_token(user_sub, refresh_token)
-    except Exception:
-        logger.exception("ensure_offline_token failed for user {}", user_sub)
