+++ mcr-core/mcr_meeting/app/infrastructure/redis.py
